% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten}
\alias{flatten}
\alias{flatten_entities}
\alias{flatten_sentences}
\alias{flatten_sentiment}
\alias{flatten_tokens}
\title{Flatten methods}
\usage{
flatten_sentences(sentences_list)

flatten_tokens(tokens_list)

flatten_entities(entities_list)

flatten_sentiment(sentiment_list)
}
\arguments{
\item{param1}{param1 description}
}
\value{
return
}
\description{
Flatten each nested JSON/list API response into a data frame
}
\examples{
example

}

