% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_config_file}
\alias{get_config_file}
\title{Fetch session-specific gcnlp default values}
\usage{
get_config_file()
}
\value{
The path to the user's \code{config_file}
}
\description{
\code{get_config_file()} gets the value of \code{config_file}
}
\examples{
\dontrun{
get_config_file()
}

}
