% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{set_api_key}
\alias{set_api_key}
\title{Set access credentials}
\usage{
set_api_key(api_key)
}
\arguments{
\item{api_key}{Your API key, from \url{https://console.cloud.google.com/apis/credentials}}
}
\value{
None
}
\description{
Define an API key
}
\examples{
\dontrun{
set_api_key("YOUR_API_KEY")
}

}

