% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{configure_googlenlp}
\alias{configure_googlenlp}
\title{Configure your computer or a server to connect to the Google Cloud Natural Language API via R functions}
\usage{
configure_googlenlp()
}
\value{
None
}
\description{
Creates variables in your .Renviron file for use by other googlenlp functions.
This will edit your .Renviron file only if you call this function directly. If you prefer not to change your .Renviron file, use the \code{set_api_key} function instead.
}
\examples{
\dontrun{
configure_googlenlp()
}

}
